requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
apipkg==3.0.1
atomicwrites==1.4.1
attrs==22.2.0
Cerberus==1.3.4
certifi==2022.12.7
chardet==5.1.0
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
execnet==1.9.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
more-itertools==9.0.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
py==1.11.0
pyee==9.0.4
pyparsing==3.0.9
pytest==7.2.1
pytest-base-url==2.0.0
pytest-forked==1.6.0
pytest-rerunfailures==11.1
pytest-xdist==3.2.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
toml==0.10.2
typing-extensions==4.5.0
urllib3==1.26.14
webdriver-manager==3.8.5